Rework the server's handler wrapper to make it more flexible
instead of taking a params map and a bodyFunc, the wrapped functions now accept a request struct which embeds the raw request and response and the mux vars.
The struct also defines a parseBody function whiuch parses a json body and a rawBody function which returns the request's raw payload.