Replace links with Docker 1.9 bridge networks
Bazooka used to rely on links to work:

* The DB container was linked into the server container
* The server container was linked into the web container

This PR proposes to replace these links (which are to be considered a
deprecated and legacy feature in Docker) with the usage of bridge
networks:

`bzk service start` will create a `bzk_net` network if it doesn't
exist, and attach the db, server and containers to it for them to be
able to communicate.

The build flow was also modified so that the various build containers
(orchestration, parser, ...) are attached to the same network when
started.

This PR alos adds a working docker compose config capable of starting
bazooka using docker compose.